Rename Editor to FileEditor to match FileManager
