+++ resources/js/store/modules/FileEditor.js
